refactor: Enumeration field arrays file structure
Shifted around the file structure for the enumeration field array compoenents so that they can be more easily readable within the repository
