+++ src/components/RulesetFormSections/EnumerationFieldArray/EnumerationTextualFieldArray/index.js
